pyproject.toml
# SPDX-License-Identifier: Apache-2.0

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "chiltepin"
version = "0.1.4"
dependencies = [
  "academy-py",
  "globus-compute-sdk>=4.3.0,<4.7.0",
  "globus-compute-endpoint>=4.3.0,<4.7.0; platform_system == 'Linux'",
  "globus-sdk>=3.65.0",
  "parsl>=2025.12.1",
  "psutil<6.0",
  "pyyaml>=6.0",
]
requires-python = ">=3.10.0"
authors = [
  {name = "Christopher Harrop", email = "Christopher.W.Harrop@noaa.gov"}
]
maintainers = [
  {name = "Christopher Harrop", email = "Christopher.W.Harrop@noaa.gov"}
]
description = "Federated NWP Workflow Tools"
readme = "README.md"
license = "Apache-2.0"
keywords = ["federated", "workflow", "parsl", "globus-compute", "hpc", "scientific-computing"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS :: MacOS X",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering",
  "Topic :: System :: Distributed Computing",
]

[project.urls]
Homepage = "https://github.com/NOAA-GSL/chiltepin"
Documentation = "https://chiltepin.readthedocs.io"
Repository = "https://github.com/NOAA-GSL/chiltepin"
"Source Code" = "https://github.com/NOAA-GSL/chiltepin"
Issues = "https://github.com/NOAA-GSL/chiltepin/issues"
"Bug Tracker" = "https://github.com/NOAA-GSL/chiltepin/issues"
Changelog = "https://github.com/NOAA-GSL/chiltepin/releases"

[project.optional-dependencies]
test = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=7.0.0",
    "ruff==0.15.4",
]
docs = [
    "sphinx==7.1.2",
    "sphinx-rtd-theme==1.3.0rc1",
]
release = [
    "build~=1.4.0",
    "twine~=6.2.0",
    "toml; python_version < '3.11'",
]

[project.scripts]
chiltepin = "chiltepin.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

[tool.coverage.run]
source = ["src/chiltepin"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.chiltepin/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
# Fail if coverage is below this threshold
fail_under = 100

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
target-version = "py310"
lint.extend-select = ["I"]
